Cleaning up server.ts and moving more auth stuff out
