+++ frontend/src/server.ts
@@ -7,35 +7,10 @@ import {
 import express from 'express';
 import { randomUUID } from 'node:crypto';
 import { join } from 'node:path';
-import {
-  authErrorResponse,
-  authErrorStatus,
-  buildLoginUrl,
-  buildLogoutUrl,
-  callbackUri,
-  changePassword,
-  clearOAuthStateCookie,
-  clearSessionCookie,
-  confirmPasswordReset,
-  confirmSignUp,
-  createOpaqueState,
-  createPkcePair,
-  authenticateWithPassword,
-  exchangeCodeForTokens,
-  getUserFromAccessToken,
-  getRequestOrigin,
-  readAuthConfig,
-  readOAuthState,
-  readSession,
-  refreshSession,
-  requestPasswordReset,
-  sessionNeedsRefresh,
-  signUpWithPassword,
-  writeOAuthStateCookie,
-  writeSessionCookie,
-  type AuthConfig,
-  type SessionPayload,
-} from './server/auth/auth-session';
+import { createAuthRouter } from './server/auth/auth.controller';
+import { resolveValidSession } from './server/auth/auth-session-resolver';
+import { readAuthConfig } from './server/auth/auth-session';
+import { formatErrorForLogSafe, registerCompactProcessErrorLogging } from './server/server-log';
 import { resolveBackendOrigin } from './backend-origin';
 
 const browserDistFolder = join(import.meta.dirname, '../browser');
@@ -56,11 +31,9 @@ const app = express();
 const angularApp = new AngularNodeAppEngine();
 const authConfig = readAuthConfig();
 
-app.use('/auth', express.json({ limit: '16kb' }));
+app.use('/auth', createAuthRouter(authConfig));
 
 const requestIdHeader = 'x-request-id';
-const maxErrorMessageLength = 256;
-const maxStackLines = 6;
 
 registerCompactProcessErrorLogging();
 
@@ -72,318 +45,6 @@ function readRequestBody(req: express.Request): Promise<Buffer> {
     req.on('error', reject);
   });
 }
-
-app.get('/auth/login', (req, res) => {
-  if (!authConfig) {
-    res
-      .status(503)
-      .json({ error: 'Authentication is not configured', code: 'auth_not_configured' });
-    return;
-  }
-  const state = createOpaqueState();
-  const pkce = createPkcePair();
-  const returnTo = sanitizeReturnTo(req.query['returnTo']);
-  const screenHint = readQueryParam(req.query['mode']) === 'signup' ? 'signup' : undefined;
-  writeOAuthStateCookie(
-    res,
-    req,
-    {
-      state,
-      codeVerifier: pkce.verifier,
-      returnTo,
-    },
-    authConfig.sessionSecret,
-  );
-  res.redirect(
-    buildLoginUrl({
-      config: authConfig,
-      redirectUri: callbackUri(req),
-      state,
-      codeChallenge: pkce.challenge,
-      screenHint,
-    }),
-  );
-});
-
-app.post('/auth/login', async (req, res) => {
-  if (!authConfig) {
-    res
-      .status(503)
-      .json({ error: 'Authentication is not configured', code: 'auth_not_configured' });
-    return;
-  }
-  const credentials = readCredentials(req.body);
-  if (!credentials) {
-    res.status(400).json({ error: 'Email and password are required', code: 'invalid_request' });
-    return;
-  }
-
-  try {
-    const result = await authenticateWithPassword({
-      config: authConfig,
-      email: credentials.email,
-      password: credentials.password,
-    });
-    if (result.status !== 'authenticated') {
-      res.status(409).json({
-        error: `Unsupported authentication challenge: ${result.challengeName}`,
-        code: 'unsupported_challenge',
-      });
-      return;
-    }
-    writeSessionCookie(res, req, result.session, authConfig.sessionSecret);
-    res.json({ authenticated: true });
-  } catch (error) {
-    console.error(`Password login failed ${formatErrorForLogSafe(error)}`);
-    res.status(authErrorStatus(error, 401)).json(
-      authErrorResponse(error, {
-        message: 'Invalid email or password',
-        code: 'invalid_credentials',
-      }),
-    );
-  }
-});
-
-app.post('/auth/signup', async (req, res) => {
-  if (!authConfig) {
-    res
-      .status(503)
-      .json({ error: 'Authentication is not configured', code: 'auth_not_configured' });
-    return;
-  }
-  const credentials = readCredentials(req.body);
-  if (!credentials) {
-    res.status(400).json({ error: 'Email and password are required', code: 'invalid_request' });
-    return;
-  }
-
-  try {
-    const result = await signUpWithPassword({
-      config: authConfig,
-      email: credentials.email,
-      password: credentials.password,
-    });
-    res.status(201).json({
-      confirmed: result.confirmed,
-      email: credentials.email,
-    });
-  } catch (error) {
-    console.error(`Password signup failed ${formatErrorForLogSafe(error)}`);
-    res.status(authErrorStatus(error, 400)).json(
-      authErrorResponse(error, {
-        message: 'Authentication request failed',
-        code: 'cognito_request_failed',
-      }),
-    );
-  }
-});
-
-app.post('/auth/confirm', async (req, res) => {
-  if (!authConfig) {
-    res
-      .status(503)
-      .json({ error: 'Authentication is not configured', code: 'auth_not_configured' });
-    return;
-  }
-  const email = readBodyString(req.body, 'email')?.trim().toLowerCase();
-  const code = readBodyString(req.body, 'code')?.trim();
-  if (!email || !code) {
-    res
-      .status(400)
-      .json({ error: 'Email and confirmation code are required', code: 'invalid_request' });
-    return;
-  }
-
-  try {
-    await confirmSignUp({
-      config: authConfig,
-      email,
-      code,
-    });
-    res.json({ confirmed: true });
-  } catch (error) {
-    console.error(`Signup confirmation failed ${formatErrorForLogSafe(error)}`);
-    res.status(authErrorStatus(error, 400)).json(
-      authErrorResponse(error, {
-        message: 'Authentication request failed',
-        code: 'cognito_request_failed',
-      }),
-    );
-  }
-});
-
-app.post('/auth/forgot-password', async (req, res) => {
-  if (!authConfig) {
-    res
-      .status(503)
-      .json({ error: 'Authentication is not configured', code: 'auth_not_configured' });
-    return;
-  }
-  const email = readEmail(req.body);
-  if (!email) {
-    res.status(400).json({ error: 'Email is required', code: 'invalid_request' });
-    return;
-  }
-
-  try {
-    await requestPasswordReset({
-      config: authConfig,
-      email,
-    });
-    res.json({ requested: true });
-  } catch (error) {
-    console.error(`Password reset request failed ${formatErrorForLogSafe(error)}`);
-    res.status(authErrorStatus(error, 400)).json(
-      authErrorResponse(error, {
-        message: 'Authentication request failed',
-        code: 'cognito_request_failed',
-      }),
-    );
-  }
-});
-
-app.post('/auth/reset-password', async (req, res) => {
-  if (!authConfig) {
-    res
-      .status(503)
-      .json({ error: 'Authentication is not configured', code: 'auth_not_configured' });
-    return;
-  }
-  const reset = readPasswordReset(req.body);
-  if (!reset) {
-    res.status(400).json({
-      error: 'Email, confirmation code, and password are required',
-      code: 'invalid_request',
-    });
-    return;
-  }
-
-  try {
-    await confirmPasswordReset({
-      config: authConfig,
-      email: reset.email,
-      code: reset.code,
-      password: reset.password,
-    });
-    res.json({ reset: true });
-  } catch (error) {
-    console.error(`Password reset confirmation failed ${formatErrorForLogSafe(error)}`);
-    res.status(authErrorStatus(error, 400)).json(
-      authErrorResponse(error, {
-        message: 'Authentication request failed',
-        code: 'cognito_request_failed',
-      }),
-    );
-  }
-});
-
-app.get('/auth/callback', async (req, res) => {
-  if (!authConfig) {
-    res
-      .status(503)
-      .json({ error: 'Authentication is not configured', code: 'auth_not_configured' });
-    return;
-  }
-  const code = readQueryParam(req.query['code']);
-  const state = readQueryParam(req.query['state']);
-  const oauthState = readOAuthState(req, authConfig.sessionSecret);
-  clearOAuthStateCookie(res, req);
-
-  if (!code || !state || !oauthState || oauthState.state !== state) {
-    res.status(400).json({ error: 'Invalid authentication callback', code: 'invalid_callback' });
-    return;
-  }
-
-  try {
-    const session = await exchangeCodeForTokens({
-      config: authConfig,
-      code,
-      codeVerifier: oauthState.codeVerifier,
-      redirectUri: callbackUri(req),
-    });
-    writeSessionCookie(res, req, session, authConfig.sessionSecret);
-    res.redirect(oauthState.returnTo);
-  } catch (error) {
-    console.error(`Authentication callback failed ${formatErrorForLogSafe(error)}`);
-    res.status(authErrorStatus(error, 502)).json(
-      authErrorResponse(error, {
-        message: 'Authentication failed',
-        code: 'token_exchange_failed',
-      }),
-    );
-  }
-});
-
-app.get('/auth/logout', (req, res) => {
-  clearSessionCookie(res, req);
-  if (!authConfig) {
-    res.redirect('/');
-    return;
-  }
-  res.redirect(
-    buildLogoutUrl({
-      config: authConfig,
-      logoutUri: `${getRequestOrigin(req)}/`,
-    }),
-  );
-});
-
-app.get('/auth/me', async (req, res) => {
-  const session = await resolveValidSession(req, res, authConfig);
-  if (!session) {
-    res.status(401).json({ authenticated: false });
-    return;
-  }
-  try {
-    const user = await getUserFromAccessToken({
-      config: authConfig!,
-      accessToken: session.accessToken,
-    });
-    res.json({ authenticated: true, email: user.email });
-  } catch (error) {
-    console.error(`Get current user failed ${formatErrorForLogSafe(error)}`);
-    res.status(authErrorStatus(error, 502)).json(
-      authErrorResponse(error, {
-        message: 'Unable to read current user',
-        code: 'current_user_unavailable',
-      }),
-    );
-  }
-});
-
-app.post('/auth/change-password', async (req, res) => {
-  const session = await resolveValidSession(req, res, authConfig);
-  if (!session) {
-    res.status(401).json({ error: 'Sign in to change your password', code: 'session_required' });
-    return;
-  }
-  const passwords = readPasswordChange(req.body);
-  if (!passwords) {
-    res.status(400).json({
-      error: 'Current password and new password are required',
-      code: 'invalid_request',
-    });
-    return;
-  }
-
-  try {
-    await changePassword({
-      config: authConfig!,
-      accessToken: session.accessToken,
-      previousPassword: passwords.currentPassword,
-      proposedPassword: passwords.newPassword,
-    });
-    res.json({ changed: true });
-  } catch (error) {
-    console.error(`Password change failed ${formatErrorForLogSafe(error)}`);
-    res.status(authErrorStatus(error, 400)).json(
-      authErrorResponse(error, {
-        message: 'Authentication request failed',
-        code: 'cognito_request_failed',
-      }),
-    );
-  }
-});
 
 app.use('/api', async (req, res) => {
   const proxyStart = performance.now();
@@ -489,107 +150,6 @@ app.use('/api', async (req, res) => {
     sendBadGatewayResponse(res, requestId);
   }
 });
-
-async function resolveValidSession(
-  req: express.Request,
-  res: express.Response,
-  config: AuthConfig | null,
-): Promise<SessionPayload | null> {
-  if (!config) {
-    return null;
-  }
-  const session = readSession(req, config.sessionSecret);
-  if (!session) {
-    return null;
-  }
-  if (!sessionNeedsRefresh(session)) {
-    return session;
-  }
-  try {
-    const refreshed = await refreshSession(config, session);
-    if (!refreshed) {
-      clearSessionCookie(res, req);
-      return null;
-    }
-    writeSessionCookie(res, req, refreshed, config.sessionSecret);
-    return refreshed;
-  } catch (error) {
-    console.error(`Session refresh failed ${formatErrorForLogSafe(error)}`);
-    clearSessionCookie(res, req);
-    return null;
-  }
-}
-
-function sanitizeReturnTo(value: unknown): string {
-  const raw = readQueryParam(value);
-  if (!raw || !raw.startsWith('/') || raw.startsWith('//') || raw.startsWith('/auth/')) {
-    return '/';
-  }
-  return raw;
-}
-
-function readQueryParam(value: unknown): string | null {
-  if (Array.isArray(value)) {
-    return readQueryParam(value[0]);
-  }
-  return typeof value === 'string' && value.trim() ? value : null;
-}
-
-function readCredentials(value: unknown): { email: string; password: string } | null {
-  if (!isRecord(value)) {
-    return null;
-  }
-  const email = readEmail(value);
-  const password = readBodyString(value, 'password');
-  if (!email || !password) {
-    return null;
-  }
-  return { email, password };
-}
-
-function readEmail(value: unknown): string | null {
-  if (!isRecord(value)) {
-    return null;
-  }
-  return readBodyString(value, 'email')?.trim().toLowerCase() || null;
-}
-
-function readPasswordReset(
-  value: unknown,
-): { email: string; code: string; password: string } | null {
-  if (!isRecord(value)) {
-    return null;
-  }
-  const email = readEmail(value);
-  const code = readBodyString(value, 'code')?.trim();
-  const password = readBodyString(value, 'password');
-  if (!email || !code || !password) {
-    return null;
-  }
-  return { email, code, password };
-}
-
-function readPasswordChange(
-  value: unknown,
-): { currentPassword: string; newPassword: string } | null {
-  if (!isRecord(value)) {
-    return null;
-  }
-  const currentPassword = readBodyString(value, 'currentPassword');
-  const newPassword = readBodyString(value, 'newPassword');
-  if (!currentPassword || !newPassword) {
-    return null;
-  }
-  return { currentPassword, newPassword };
-}
-
-function readBodyString(value: unknown, key: string): string | null {
-  if (!isRecord(value)) {
-    return null;
-  }
-  const item = value[key];
-  return typeof item === 'string' ? item : null;
-}
 
 function readRequestId(req: express.Request): string | null {
   const value = req.headers[requestIdHeader];
@@ -744,96 +304,6 @@ function sendBadGatewayResponse(res: express.Response, requestId: string): void 
   } catch {
     /* exhausted fallbacks */
   }
-}
-
-function formatErrorForLogSafe(error: unknown): string {
-  try {
-    return formatErrorForLog(error);
-  } catch {
-    return 'unknown error';
-  }
-}
-
-function registerCompactProcessErrorLogging(): void {
-  process.on('uncaughtException', (error) => {
-    console.error(`uncaughtException ${formatErrorForLog(error)}`);
-  });
-  process.on('unhandledRejection', (reason) => {
-    console.error(`unhandledRejection ${formatErrorForLog(reason)}`);
-  });
-}
-
-function formatErrorForLog(error: unknown): string {
-  if (error instanceof Error) {
-    const details = compactObjectDetails(error);
-    const stack = compactStack(error.stack);
-    return [formatErrorNameAndMessage(error.name, error.message), details, stack]
-      .filter(Boolean)
-      .join(' ');
-  }
-
-  if (isRecord(error)) {
-    return [
-      formatErrorNameAndMessage(readString(error, 'name'), readString(error, 'message')),
-      compactObjectDetails(error),
-    ]
-      .filter(Boolean)
-      .join(' ');
-  }
-
-  return truncate(String(error), maxErrorMessageLength);
-}
-
-function formatErrorNameAndMessage(name: string | null, message: string | null): string {
-  const errorName = name || 'Error';
-  return message ? `${errorName}: ${truncate(message, maxErrorMessageLength)}` : errorName;
-}
-
-function compactObjectDetails(error: Error | Record<string, unknown>): string {
-  const record = error as Record<string, unknown>;
-  const details = [
-    formatDetail(record, 'code'),
-    formatDetail(record, 'status'),
-    formatDetail(record, 'statusText'),
-    formatDetail(record, 'url'),
-    formatDetail(record, 'type'),
-  ].filter(Boolean);
-
-  return details.length ? `(${details.join(' ')})` : '';
-}
-
-function formatDetail(record: Record<string, unknown>, key: string): string | null {
-  const value = record[key];
-  if (typeof value === 'string' || typeof value === 'number' || typeof value === 'boolean') {
-    return `${key}=${truncate(String(value), 200)}`;
-  }
-  return null;
-}
-
-function compactStack(stack: string | undefined): string {
-  if (!stack) {
-    return '';
-  }
-  const lines = stack
-    .split('\n')
-    .slice(1, maxStackLines + 1)
-    .map((line) => line.trim())
-    .filter(Boolean);
-
-  return lines.length ? `stack="${lines.join(' | ')}"` : '';
-}
-
-function readString(record: Record<string, unknown>, key: string): string | null {
-  const value = record[key];
-  return typeof value === 'string' ? value : null;
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null;
-}
-
-function truncate(value: string, maxLength: number): string {
-  return value.length > maxLength ? `${value.slice(0, maxLength - 1)}…` : value;
 }
 
 /**
